2024/pyproject.toml
[project]
name = "aoc2024"
version = "0.1.0"
description = "Solutions to Advent of Code 2024"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.0",
    "python-dotenv>=1.0.1",
]
